requirements.txt
asgiref==3.8.1
astroid==3.1.0
black==24.4.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.8
Django==5.0.4
django-jazzmin==3.0.0
django-restframework==0.0.1
djangorestframework==3.15.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mccabe==0.7.0
mypy==1.10.0
mypy-extensions==1.0.0
openapi-codec==1.3.2
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
psycopg2-binary==2.9.9
pylint==3.1.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
sqlparse==0.5.0
tomlkit==0.12.4
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
